# Replace existing waitforcompaction with new WaitForCompact API in db_bench_tool